refactor: organize codebase
