+++ src/lib/is.ts
