+++ src/lib/private-api.ts
@@ -1,0 +1,4 @@
+export { isOneOfMultipleTypes } from './is';
+export { matchesSchema } from './schema';
+export { isMultipleOf, testNumberWithinBounds } from './number-helpers';
+export { DATATYPE } from './data-type';
