+++ src/spec/type-string.spec.ts
@@ -1,8 +1,11 @@
 import { test } from 'ava';
-import { is } from '../src/is';
-import { DataType } from '../src/data-type';
-import { matchesSchema } from '../src/schema';
-import { getDataTypeUseCases, stringPatternUseCases, validStringUseCases } from './test-cases';
+import { DataType, is } from '@lib';
+import { matchesSchema } from '@lib-private';
+import {
+  getDataTypeUseCases,
+  stringPatternUseCases,
+  validStringUseCases
+} from './test-cases/index';
 
 const currentDataType = DataType.string;
