--- tools/benchmarks/tsconfig.json
@@ -1,12 +1,0 @@
-{
-  "extends": "../../tsconfig-base.json",
-  "compilerOptions": {
-    "lib": ["es2015"],
-    "module": "commonjs",
-    "target": "es5",
-    "declaration": false,
-    "strict": true,
-    "types": ["node"],
-    "downlevelIteration": true
-  }
-}
