--- tsconfig-base.json
@@ -1,18 +1,0 @@
-{
-  "compilerOptions": {
-    "module": "es2015",
-    "target": "es2015",
-    "forceConsistentCasingInFileNames": true,
-    "importHelpers": true,
-    "moduleResolution": "node",
-    "noEmitHelpers": true,
-    "noImplicitAny": true,
-    "noImplicitReturns": true,
-    "noImplicitThis": true,
-    "noUnusedLocals": true,
-    "noUnusedParameters": true,
-    "removeComments": false,
-    "sourceMap": false,
-    "strict": true
-  }
-}
