+++ tsconfig-bench.json
@@ -1,0 +1,11 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "module": "commonjs",
+    "target": "es2015",
+    "types": ["node"]
+  },
+  "files": [
+    "src/benchmarks/index.ts"
+  ]
+}
