+++ tsconfig-lib.json
@@ -1,9 +1,10 @@
 {
   "extends": "./tsconfig.json",
   "compilerOptions": {
-    "outDir": "dist",
-    "declaration": true,
-    "emitDeclarationOnly": true
+    "module": "es2015",
+    "target": "es5",
+    "outDir": "__build",
+    "sourceMap": true
   },
   "files": [
     "src/index.ts"
