+++ tsconfig.json
@@ -1,12 +1,27 @@
 {
-  "extends": "./tsconfig-base.json",
   "compilerOptions": {
     "module": "es2015",
-    "target": "es5",
-    "outDir": "__build",
-    "sourceMap": true
-  },
-  "files": [
-    "index.ts"
-  ]
+    "target": "es2015",
+    "forceConsistentCasingInFileNames": true,
+    "importHelpers": true,
+    "moduleResolution": "node",
+    "noEmitHelpers": true,
+    "noImplicitAny": true,
+    "noImplicitReturns": true,
+    "noImplicitThis": true,
+    "noUnusedLocals": true,
+    "noUnusedParameters": true,
+    "removeComments": false,
+    "sourceMap": false,
+    "strict": true,
+    "baseUrl": "./",
+    "paths": {
+      "@lib": [
+        "src/index.ts"
+      ],
+      "@lib-private": [
+        "src/lib/private-api.ts"
+      ]
+    }
+  }
 }
